requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
djangorestframework==3.14.0
idna==3.6
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
